+++ src/drivers/sdl/lv_sdl_window.c
@@ -39,7 +39,7 @@ typedef struct {
     uint8_t * buf1;
     uint8_t * buf2;
 #endif
-    uint8_t zoom;
+    float zoom;
     uint8_t ignore_size_chg;
 } lv_sdl_window_t;
 
@@ -149,7 +149,7 @@ void lv_sdl_window_set_resizeable(lv_display_t * disp, bool value)
     SDL_SetWindowResizable(dsc->window, value);
 }
 
-void lv_sdl_window_set_zoom(lv_display_t * disp, uint8_t zoom)
+void lv_sdl_window_set_zoom(lv_display_t * disp, float zoom)
 {
     lv_sdl_window_t * dsc = lv_display_get_driver_data(disp);
     dsc->zoom = zoom;
@@ -157,7 +157,7 @@ void lv_sdl_window_set_zoom(lv_display_t * disp, uint8_t zoom)
     lv_refr_now(disp);
 }
 
-uint8_t lv_sdl_window_get_zoom(lv_display_t * disp)
+float lv_sdl_window_get_zoom(lv_display_t * disp)
 {
     lv_sdl_window_t * dsc = lv_display_get_driver_data(disp);
     return dsc->zoom;
