+++ src/drivers/sdl/lv_sdl_window.h
@@ -35,9 +35,9 @@ lv_display_t * lv_sdl_window_create(int32_t hor_res, int32_t ver_res);
 
 void lv_sdl_window_set_resizeable(lv_display_t * disp, bool value);
 
-void lv_sdl_window_set_zoom(lv_display_t * disp, uint8_t zoom);
+void lv_sdl_window_set_zoom(lv_display_t * disp, float zoom);
 
-uint8_t lv_sdl_window_get_zoom(lv_display_t * disp);
+float lv_sdl_window_get_zoom(lv_display_t * disp);
 
 lv_display_t * _lv_sdl_get_disp_from_win_id(uint32_t win_id);
